requirements.txt
ultralytics>=8.0.0
Pillow>=9.0.0
requests>=2.31.0
aiohttp>=3.8.0
streamlit>=1.32.0